cpuidle: lpm-workarounds: Skip L2 SPM enabling in HLOS
Due to a hardware issue on 8909, secure world needs to disable
and enable L2 SPM to get the proper context in secure watchdog
bite cases.

With this workaround there is a race in programming L2 SPM between
HLOS and secure world. This leads to stability issues. To avoid
this program L2 SPM enable in secure world.

Change-Id: I426289e1fc66664ed9c4ecf81c4106ce2c7c1145
Signed-off-by: Anil Kumar Mamidala <amami@codeaurora.org>
Signed-off-by: Chinkit Kumar,Kirti Kumar Parmar <parma@codeaurora.org>