cnss2: Consolidate logging message
Currently the code makes liberal use of macros that print a log
message (with pr_err or pr_debug) and then passes the same string
on to the IPC logging mechanism. The problem is that it doesn't
actually end up being the same string in the binary. Using pr_err
or one of its friends appends the KERN_* code to the front of the
string with the pre-processor and the IPC logger just uses the
passed in string. Every string used by the macros ends up appearing
twice in the binary, once with KERN_* prepended and the other not.

This change fixes this duplication issue by appending KERN_* to the
front of the IPC logger for CNSS2 driver.

Change-Id: I62facd7aaf23bd06fad84d272144f1a50ad539a5
Signed-off-by: Yue Ma <yuem@codeaurora.org>