soc: qcom: Adjust policy min at cluster PC enter and exit
Set the policy min of big cluster to lowest frequency at
cluster power collapse enter and set it to user supplied
frequency at power collapse exit.  This is to allow the
voltage in case of common apps rail to go down to the small
cluster operating level when big cluster is not in use.

This driver supports below module parameters.

min_freq_adjust: Enables driver functionality. It takes
only [yY1] as input. Does not allow to disable once enabled.
Default is disabled.

min_freq_floor: The floor of policy min to be voted when big
cluster is out of power collapse. Default is 1094400.

min_down_delay_ms: Latency before the driver votes down
the floor of policy min to lowest corner.  Default is 80ms.

Change-Id: I4ede7730953490b99a5b8072948c0371cc1b9b8e
Signed-off-by: Ramakrishna Gottimukkula <rgottimu@codeaurora.org>