ARM: dts: msm: Do not initialize the reset line for sda845+sdx24
In the event of the modem error fatal, it excepts the APQ to hold
the reset line as it is during its bootup. Initializing this to
LOW would trigger a cold reset on the modem. As a result, the
modem is powered-off, and the ramdumps are lost. Avoiding
initializing the reset gpio to LOW would hold the modem up
(unchanged) and the ramdumps are available.

Change-Id: I7e1c50583d74cbe9c113086940e78c7ad7901ec4
Signed-off-by: Raghavendra Rao Ananta <rananta@codeaurora.org>